Service Layer, Repository and more
Created service layer (JournalService, RevisionService) in order to separate concerns and follow MVC principles.

Added Repositories to handle all DB functions
