+++ src/main/java/com/selekode/service/StatsService.java
@@ -1,0 +1,9 @@
+package com.selekode.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class StatsService {
+	
+
+}
