+++ src/main/resources/templates/about.html
@@ -88,15 +88,9 @@
 
         <!-- Buttons Section -->
         <div class="space-y-4 w-full flex flex-col items-center">
-            <a href="/journal/addEntry">
-                <button class="w-12 h-12 rounded-full bg-[#23A3BB] font-bold text-white flex items-center justify-center hover:bg-[#1A8F9A] mb-4 transition-colors duration-300">
-                    +
-                </button>
-            </a>
-
             <!-- Rounded buttons for other actions -->
             <a href="/journal/load" class="block w-full">
-                <button class="w-full text-left py-2 px-4 bg-[#1A8F9A] rounded-full transition-colors duration-300 hover:bg-[#23A3BB]">
+                <button class="w-full text-left py-2 px-4 bg-[#23A3BB] rounded-full transition-colors duration-300 hover:bg-[#1A8F9A]">
                     Diario
                 </button>
             </a>
@@ -111,7 +105,7 @@
                 </button>
             </a>
             <a href="/about" class="block w-full">
-                <button class="w-full text-left py-2 px-4 bg-[#23A3BB] rounded-full transition-colors duration-300 hover:bg-[#1A8F9A]">
+                <button class="w-full text-left py-2 px-4 bg-[#1A8F9A] rounded-full transition-colors duration-300 hover:bg-[#23A3BB]">
                     Sobe Topaz
                 </button>
             </a>
@@ -122,45 +116,14 @@
     <!-- Main content -->
     <div class="main-content  flex-grow bg-[#0F2529] p-6">
         <main class="container mx-auto mt-6 p-4">
-            <h1 class="text-2xl font-bold mb-4 text-white">Diario</h1>
+            <h1 class="text-2xl font-bold mb-4 text-white">Sobre Topaz</h1>
 
-            <!-- Iterate over journal entries -->
-            <div th:each="journalEntry : ${journalEntries}" x-data="{ open: false }" class="relative bg-[#172E38] p-6 rounded-lg shadow-2xl mb-6">
-                <!-- Buttons Positioned in the Top Right -->
-                <div class="absolute top-4 right-4 flex space-x-2">
-                    <!-- Show/Hide Button -->
-                    <button @click="open = !open" 
-                            class="px-4 py-2 rounded-full bg-[#23A3BB] text-white text-sm transition hover:bg-[#1A8F9A]">
-                        <span x-text="open ? 'Ocultar' : 'Mostrar'"></span>
-                    </button>
-            
-                    <!-- Edit button -->
-                    <a th:href="@{/journal/editEntry/{id}(id=${journalEntry.id})}" 
-                       class="px-4 py-2 rounded-full bg-[#23A3BB] text-white text-sm transition hover:bg-[#1A8F9A]">
-                        Editar
-                    </a>
-                    
-                    <!-- Delete button -->
-                    <button th:attr="onclick='openModal(' + ${journalEntry.id} + ')'" 
-                            class="px-4 py-2 rounded-full bg-[#E74C3C] text-white text-sm transition hover:bg-[#D63C2D]">
-                        Eliminar
-                    </button>
-                </div>
-            
-                <!-- Journal Entry Content -->
-                <div class="mr-4">
-                    <!-- Date -->
-                    <p class="text-lg italic font-bold mb-2" style="color: #23A3BB;" th:text="${journalEntry.date}"></p>
-            
-                    <!-- Content to Show/Hide -->
-                    <div x-show="open" x-collapse>
-                        <hr class="w-full border-t-[1px] border-[#1A8F9A] mb-4 mt-4">
-
-                        <h2 class="text-2xl font-bold mb-4 text-white" th:text="${journalEntry.title}"></h2>
-                        <p class="text-base mb-4 text-white" th:text="${journalEntry.content}"></p>
-                    </div>
-                </div>
+            <div class="bg-[#172E38] p-6 rounded-lg shadow-2xl">
+                <h1 class="text-l mb-2" style="color: #23A3BB;">Versión 0.0.2</h1>
+                <h1 class="text-xl mb-2  text-white">Topaz es un diario y rastreador de emociones</h1>
             </div>
+        </main>
+    </div>
             
             <!-- Alpine.js -->
             <script src="https://cdn.jsdelivr.net/npm/alpinejs@3.x.x/dist/cdn.min.js" defer></script>
