+++ src/main/resources/templates/journal_addEntry.html
@@ -91,12 +91,16 @@
                     Revisión
                 </button>
             </a>
-            <button class="w-full text-left py-2 px-4 bg-[#23A3BB] rounded-full transition-colors duration-300 hover:bg-[#1A8F9A]">
-                Stats
-            </button>
-            <button class="w-full text-left py-2 px-4 bg-[#23A3BB] rounded-full transition-colors duration-300 hover:bg-[#1A8F9A]">
-                Sobre Topaz
-            </button>
+            <a href="/stats/load" class="block w-full">
+                <button class="w-full text-left py-2 px-4 bg-[#23A3BB] rounded-full transition-colors duration-300 hover:bg-[#1A8F9A]">
+                    Estadísticas
+                </button>
+            </a>
+            <a href="/about" class="block w-full">
+                <button class="w-full text-left py-2 px-4 bg-[#23A3BB] rounded-full transition-colors duration-300 hover:bg-[#1A8F9A]">
+                    Sobe Topaz
+                </button>
+            </a>
         </div>
     </div>
